requirements-dev.txt
pytest==8.3.1
pytest-django==4.8.0
model-bakery==1.18.2
locust==2.29.1

